Replace SimpleArrayMap usage with HashMap in RenderState
Summary: `SimpleArrayMap/ArrayMap` is not safe to use in multithreaded environment as it is not designed to do so in the first place. Moreover, `SimpleArrayMap` uses static internal cache for its backing array and once it gets corrupted in one usage in multithreaded environment, it affects ALL other current usages even if some of other usages are legit (accessing it only from single thread). As we cannot guarantee that no `SimpleArrayMap` will be used in our app, safest way is to remove all current usages. Replacing it with `HashMap` in `RenderState`

Reviewed By: IanChilds

Differential Revision: D9013687

fbshipit-source-id: 780aac0cea4949f07b1e647be2eeb5da62abf84a