Replace SimpleArrayMap/ArrayMap usages with HashMap in Working range
Summary: `SimpleArrayMap` is not safe to use in multithreaded environment as it is not designed to do so in the first place. Moreover, `SimpleArrayMap` uses static internal cache for its backing array and once it gets corrupted in one usage in multithreaded environment, it affects ALL other current usages even if some of other usages are legit (accessing it only from single thread). As we cannot guarantee that no `SimpleArrayMap` will be used in our app, safest way is to remove all current usages. Replacing it with `HashMap` in working range.

Reviewed By: topwu

Differential Revision: D9013708

fbshipit-source-id: 14abbfb11ccd0576430dcce87fc5e3c871699f2b