Replace SimpleArrayMap usages with HashMap in dataflow package
Summary: `SimpleArrayMap` is not safe to use in multithreaded environment as it is not designed to do so in the first place. Moreover, `SimpleArrayMap` uses static internal cache for its backing array and once it gets corrupted in one usage in multithreaded environment, it affects ALL other current usages even if some of other usages are legit (accessing it only from single thread). As we cannot guarantee that no `SimpleArrayMap` will be used in our app, safest way is to remove all current usages. Replacing it with `HashMap` where insertion order does not matter and with `LinkedHashMap` where it does matter in dataflow package.

Reviewed By: IanChilds

Differential Revision: D9013797

fbshipit-source-id: 298cc214e156a452d552849b1bccf2554f05df3a