Cache and share LayoutState computation between threads (#385)
Summary:
This mitigates any duplicate layout that happens between the main and background threads while also caching the value as a result on a Future.

The existing behavior can be simplified to the following:
- main thread -> measure() -> **calculateLayoutState()** -> mMainThreadLayoutState
- bg thread -> calculateLayout() -> **calculateLayoutState()** -> mBackgroundLayoutState
- **calculateLayoutState()** will then call the static function LayoutState.calculate() which eventually leads to calling the Components' OnCreateLayout functions down the tree.
- Eventually the mBackgroundLayoutState is copied over to mMainThreadLayoutState through backgroundLayoutStateUpdated() ->  setBestMainThreadLayoutAndReturnOldLayout()

Obviously a lot of details I'm leaving out, but that's the general gist. The important note here is that there are many subtleties in the code between the main and bg functions, intentional or not, that we don't want to inadvertently affect.

The issue we're seeing is that on initial layout calculateLayoutState() is often called twice for the same layout because the background threads don't finish calculating layouts and report them back to the main thread in time before the measure() function is called. Ironically, this seems exacerbated by expensive OnCreateLayout calls, which are exactly the calls we don't want to duplicate during initial layout.

This change will share a FutureTask between main and bg threads that calculate the same layout (determined by LayoutStateFuture.equals()). That LayoutState will be also be cached on the Future until the next run with different params.
Pull Request resolved: https://github.com/facebook/litho/pull/385

Reviewed By: mihaelao

Differential Revision: D8731943

Pulled By: astreet

fbshipit-source-id: 2489f206c4916e37fd96b98e9bab9f5962545e91