Correct range computation order
Summary:
When I was poking around how Litho renders our thread view, I noticed that items outside viewport but within the range are computed on background thread in wrong order.

I tested this using a thread with messages of "message 1, message 2, message 3, ... message n", with "message 1" being the earliest at top and "message n" being the latest at bottom.

Let's say we load 20 messages (1...20), and 15 of them will be visible in viewport.

Before this change, we compute message 1 to 5 (outside viewport) from 1 to 5, but because our thread view grows/stacks from bottom, 5 will appear next as users scroll up.

Because there is only one layout thread and it is serial, the order is quite important IMO.

This diff fixes the order by detecting if the layout is reversed.

Reviewed By: muraziz

Differential Revision: D9068549

fbshipit-source-id: c9b9da2fef6ffcfec0cf13b4897897e2bbbec4da