Extract range traversing logic to another class
Summary:
`RecyclerBinder` is not very easy to test, so I extracted the range traversing logic to another class.

As I worked on this, I realized that we don't need to know about `stackFromEnd` or `reverseLayout`, the better heuristic would be to **start from the middle of the visible range and expand outward**, that way items adjacent to the visible range (more likely to appear next) will be computed before items that are farther away.

Reviewed By: muraziz

Differential Revision: D9117515

fbshipit-source-id: 07e3e205dd7ba6c1a537c3e7b706e0a7ea4806b5