Drop log events without log tag
Summary:
LogTreePopulator now returns the event it modifies and can `null` it.
That allows us to put further constraints in here - before that, we should
probably rename it though - like dropping events unless a context chain is
present.

For now, however, this is only to increase the signal in our logs.

Reviewed By: IanChilds

Differential Revision: D9179016

fbshipit-source-id: 2b6ca0ecf0b7a263cee38d73e73c3f9e5be795fb