Set up onPreDraw listener in IncrementalMountHelper
Summary:
This has the basic setup for a new experiment to mount in onPreDraw. This *should* ensure that we *always* mount, catching the small number of cases we have right now where mounting is not successful (e.g. in a ScrollView).

Future diffs will add some more code around this - turning off all other ways to mount when this is true, and an experiment to see whether this is accurate and performant.

Reviewed By: astreet

Differential Revision: D9179474

fbshipit-source-id: 16644e998daa141517ecb7a934a1ccc266e90f63