Implement fastpath for DataDiffSection
Summary:
We noticed that changeset generation for certain use cases can be quite slow, taking around 50ms. After looking into this, I found that `DiffUtil` sucks at handling two completely different list (probably the algorithm's worst case scenario).

This diff adds an optional data identifier prop to `DataDiffSectionSpec` as a way to inform the section if it's still the same list or not, if not, diffing should be skipped and a straight replacement changeset be used instead.

Reviewed By: muraziz

Differential Revision: D9237127

fbshipit-source-id: 5b2cada271c8abff188b628fbbdd8dd0ac85e3d8