Add optional prop bottom shadow margin to card spec
Summary: Card Spec adds the content to be rendered as a child of a column to which bottom margin is applied to. The bottom margin is calculated by multiplying the elevation dip with a shadow multiplier constant, and this value changes depending on the screen size.This produces inconsistent results for my full bleed thumbnail, as sometimes the bottom margin is too great and create a space between the thumbnail and the preview box. Passing in the value of the bottom margin as an optional prop allows the thumbnail to be full bleed for all devices.

Differential Revision: D9257446

fbshipit-source-id: e6cd0387442b143138b90e447d42916721ef93e5