Cancel dropped log events
Summary:
APIs may require that events are properly released instead of just dropped.
I recently added some restrictions to when MOUNT events, in particular, are
logged and this led to a bunch of events never getting properly canceled/released.

This adds the necessary interface to the ComponentsLogger and implements it
where necessary.

Reviewed By: IanChilds

Differential Revision: D9447798

fbshipit-source-id: 106191d9aabf0ccc8db098ae8edc93e3386d795a