Track crashes in stacktraces across threads.
Summary:
In Litho we have instances where events or stateUpdates triggers runnables on other threads. If this code crashes we have not much visibility on the stacktrace that actually originated that. This diff introduces ThreadTracingRunnable to overcome this issue.
ThreadTracingRunnable tries to preserve the stacktrace of the codeflow across threads to show the actual origin of a crash.

Reviewed By: mihaelao

Differential Revision: D8916056

fbshipit-source-id: 8c826fbe0cb7ec8768db130d3e5b918e8f26d7e2