Print hierarchy stacktrace on createLayout crashes
Summary:
If a crash happens in a Litho widget or a Component re-used in multiple places, at the moment the stacktrace doesn't really point to the origin of the crash.
With this diff, we print in the stacktrace, the hierarchy of Components involved into that crash to have a better picture of what it crashing.

Reviewed By: IanChilds

Differential Revision: D8915950

fbshipit-source-id: 3b4c1a921981be25a7a305fdb0f2d85bc981ef9f