Add API to request scroll to position that includes snapping
Summary: `RecyclerEventsController#requestScrollToPosition(int, boolean)` scrolls to the target position until it is fully visible. However, when `RecyclerCollectionComponent` uses non-default snapping like snap_center it might be desirable to scroll to the target position in such way that target will be visible at the center (or wherever it is snapped). Added `requestScrollToPositionWithSnap` API for that purpose.

Reviewed By: IanChilds

Differential Revision: D9455834

fbshipit-source-id: df171c7a4f7ac4291b40cd5065fa92525cf0a56e