Add open source Flipper support
Summary:
We can finally merge the fb-only and oss versions of the sample app
as Flipper is open source. This also means you can use Flipper +
Instant Run (which comes in quite handy for me).

The open-source BUCK build continues to *build*, but doesn't
actually *run*. That's a trade-off we've had for a while
without too many complaints. If this bothers you, dear reader of this message,
pull requests are welcome.

The controller you requested could not be found.

Reviewed By: IanChilds

Differential Revision: D9528697

fbshipit-source-id: 59602cfe3b85478f8df31b5a5ccbd84ecb280bba