Async init range based on viewport size
Summary:
proposal for getting rid of the synchronous layout triggered by init range. The synchronous call is necessary to get the size of the first item and based on that to estimate how many items fit in the viewport.
This triggers a few (MIN_RANGE_SIZE) layouts asynchronously and when those layouts complete, it keeps scheduling the asynchronous layouts until enough items to fill the viewport complete. When that happens, that number is used as the estimated viewport count, and the first part of the range is already completed. If a compute range was requested in measure() but the range was not initialized yet, that compute range call will be completed now too.

If not enough items are inserted in the binder to fill the viewport, the estimated viewport count will be marked as such, so that if more items are inserted later the range can be adjusted.

The next step would be to adjust the range size dynamically as we compute range.

Depends on D9420887

Reviewed By: astreet

Differential Revision: D9420904

fbshipit-source-id: 92d70ebf18d9957693979f8b67175cf2f937b0cb