Fixes the issue where updates to common props are not applied to a view when it's measure does not change
Summary:
When updating the MountState of a MountSpec the method `MountState#updateMountItemIfNeeded(...)` does not diff common view properties hosted in `NodeInfo` and `ViewNodeInfo`.

In order to fix this we implemented a `isEquivalentTo()` method for `NodeInfo` and `ViewNodeInfo`. These methods check if 2 node info have equivalent properties. This enable us to diff the node info on the current mount item and the new layout output.

This feature is behind an experiment which is disabled by default. To enable it set `ComponentsConfiguration#enableViewInfoDiffingForMountStateUpdates` to `true`.

Reviewed By: IanChilds

Differential Revision: D9496989

fbshipit-source-id: 0e51d73e9b53719c4a427004a14efa3bf46206fb