Visibility event for visibility rect changed
Summary:
for implementing a focused visibile functionality on the section level, we need to know how much a component is visible on screen to know how much of the screen the entire parent section is covering. This adds a new event that is dispatched when the visible rect of the component changes and contains the full size and visible size of that component.
If this and the next diff make this overly complicated, happy to discuss other approaches.

Reviewed By: IanChilds

Differential Revision: D9558098

fbshipit-source-id: fbb3b0feda9849e9d24c7d64113b6f213003a85b