Add ChangeSetStats
Summary:
This sets up ChangeSetStats, similar to MountStats, that we then can log with ComponentsLogger. The structure is only allocated if we trace a given `PerfEvent`.

This diff only sets up the data structure and the necessary threading of the `enableStats` flag. The actual wiring to ComponentsLogger happens in the next diff.

Reviewed By: mihaelao

Differential Revision: D9607835

fbshipit-source-id: 4734f5c03d474a25446e92cf996afd9ad974daf3