Add experiment for unmounting when not visible
Summary: Previously we would just early-return when we were totally off-screen. However, the problem with this is that means that we don't unmount things if we move straight off the screen, which means that visibility events are not correctly fired. Instead, we should mount with an empty rect, and we'll also return early if we're mounting again with another empty rect to avoid doing too much work.

Reviewed By: muraziz

Differential Revision: D9772208

fbshipit-source-id: fc7f16c94b785aa9dd863ab3adeef9afcd4c43f7