add key to duplicate key log message
Summary:
This makes this exception easier to debug.

FB: @[591189524:Shenwei] and I ran into a production trace with this bug and were unsure how it came about. Having the actual key in the exception trace would it make it easier to know where it's coming from. My only concern is will errors still be aggregated if we vary the exception by key? Putting this out for comments.

Reviewed By: Daniel9z

Differential Revision: D9784824

fbshipit-source-id: feeaf3d9a3d5c776c70524b065011da3f5b9ae9b