make REQUIRED_PROP_NAMES/COUNT variables static
Summary:
These are generated as instance variables in every Builder class when they could be static. This diff updates the codegen to do that.

FB: From the NT MobileLab tests, it looks like avoiding the per builder allocation doesn't have an significant effect on NT's payload, but it certainly can't hurt, and it might help more on components with more props. It also probably doesn't matter for Builders that Redex is inlining, but once again doesn't hurt.

Reviewed By: IanChilds

Differential Revision: D9785952

fbshipit-source-id: 5292b51b7021c9ee9754c371b6574d1fb9cfcc72