Unmount sticky header when hiding it
Summary: Sticky header is implemented in a such way that same ComponentTrees are shared between two different views (LithoView from RecyclerView and LithoView used as overlay sticky header). Unless we explicitly unmount lithoview of overlay sticky header when hiding it, next time it needs to be displayed it will think that it is already mounted (due to RecyclerBinder#mMountedView and will skip mounted), but in fact has been unmounted via onViewRecycler.

Reviewed By: IanChilds

Differential Revision: D9923869

fbshipit-source-id: 8bdbbd50409af24e8389c8153493864342170464