Disable asyncInitRange if canMeasure is enabled
Summary:
with async init range, when canMeasure is enabled and the recycler might try to measure before the first layout could finish so the RecyclerView will measured with height 0. When canMeasure and asyncInitRange are enabled, this change will always compute the first layout synchronously during measure, and immediately after getting these measurements start the computeRange calculation.
Depends on D10051607

Reviewed By: astreet

Differential Revision: D10083834

fbshipit-source-id: 70daf4192839a83c7117d19d32802cf736ef0f3d