Add configuration for HScroll async mode
Summary: We're introducing an HScroll async mode for hscrolls embedded in vertical scrolls where we want to make sure the hscroll has its content created before it appears on screen (so that we don't block the main thread). The full implementation of this parameter is coming in the next diff.

Reviewed By: mihaelao

Differential Revision: D10103157

fbshipit-source-id: 72ef9f53fbab7f34f78b73ee4e90debdd0bf1d4d