Checks around Component not being null in LayoutOutput and MountItem
Summary:
Looks like we have some crashes where a MountItem has a null Component when we try to unmount it. This should never happen.
With this commit we're trying to understand if/when a null Component is set to a LayoutOutput or MountItem.
Another very bad possibility is that MountState is called by another Thread operating on the same list of MountItem with obvious sync issue.

Reviewed By: mihaelao

Differential Revision: D10199972

fbshipit-source-id: 883611ee1007f00175c6133a7c0619a7e4ad2f1f