remove DLDrawable reference from MountItem
Summary:
A part of DisplayList revamping campaign (5/n): MountItem used to contain a reference to DL, which would lead to the code like
   Drawable contentDrawable = mountItem.getContent();
   Drawable drawable = mountItem.getDisplayListDrawable() == null ? contentDrawable : mountItem.getDisplayListDrawable();
in multiple places, which is unwanted, and leads to unwanted hard to debug behavior if the check isn't done where it should be. We'll be introducing a more flexible solution in the following diff.

Reviewed By: astreet

Differential Revision: D10161977

fbshipit-source-id: 63a8bf21ae24fc91d38ec76b55b35a1828cd143c