Add global stats counters
Summary:
This adds a new global counter for Litho operations. While the absolute numbers may provide limited use over the runtime of an application, taking snapshots can be helpful for regression analysis, especially for startup scenarios.

I'm using atomic longs for the counter(s) as the cases of thread contention should be very limited and cost of incrementing in uncontended scenarios is very low. With general availability of Java 8 on Android, we may want to consider switching to [LongAdder](https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/atomic/LongAdder.html)s as they are cheaper if a CAS fails due to contention by keeping the difference in a thread-local.

The set of available counters will grow over time. State updates is just the first to provide some parity with ComponentKit.

Reviewed By: astreet

Differential Revision: D9973149

fbshipit-source-id: ff1c5b1bb9f5573458d8fda70bad70bd4c8280a9