Use Context directly for acquiring and releasing mount content
Summary: This is a small step towards having ComponentContext not extend Context. In particular, we don't need a ComponentContext when we are acquiring and releasing mount content - we can just use Context directly, which is what we do here. To avoid the leaks that we saw before with ComponentContext being used, we crash if that is the case. There are a limited number of controlled callsites so this shouldn't lead to crashes in prod.

Reviewed By: astreet

Differential Revision: D10239868

fbshipit-source-id: ab10afef4992c575ef5e3a9e74584190da77580e