change short to int
Summary:
> **TLDR;** Get rid of the following nit unless we know that this optimization is needed.

{F141616595}

Using short with `IntDef` throws lint errors in Android. In addition, using `short` instead `int` makes *little* difference in terms of memory used. Object fields in JVM (should be the same in Dalvik & ART) use offsets that are multiples of a 32 bits. So when we declare a primitive member or local variable (say a short) it will be stored in a 32 bit cell, just like an int.

> exceptions: `long`, `double` and `arrays` of primitive types.

Reviewed By: muraziz

Differential Revision: D10446045

fbshipit-source-id: 91a769e1c47028b00c1e066c3eac32fa74a3f51e