fix selecting a position before mounting a circular recycler and it either resets to middle or does not allow user to scroll to the left
Summary: Right now in mount for circular recycler binder, either there is a preselected position, in which case the binder will scroll to that position, or it will check that it is circular, and scroll to the 0th element after jumping to the middle.  Thus, there is no way to select an item in circular mode, and scroll it to that element after jumping to the middle.  This diff plans to solve that.

Reviewed By: astreet

Differential Revision: D10361960

fbshipit-source-id: 5c5c982495130f0a835ef4eee0edd126b4381fe3