Sections should allow differently-typed Sections at their root
Summary: Previously, if you gave two different kinds of Sections to your SectionTree, it would try to cast one to the other and that would lead to a crash. We handle that by checking if we have the same type, and if we don't we just remove everything from the current root and add everything from the new root. We also have to add a check in `createNewTreeAndApplyStateUpdates` because that also assumes that we have the same type - now that we don't, we might not have children for the current root when the nextRoot is not a DiffSectionSpec (since the current root could be a DiffSectionSpec).

Reviewed By: astreet

Differential Revision: D10299462

fbshipit-source-id: e7f5bc15c668fb5b113ba7f686daeeda2cef42c7