Use ViewCreator's hashcode instead of object itself as key to the component pool
Summary: Using `ViewCreator` object itself as the key to components pool whose lifecycle is the lifecycle of the whole app might be dangerous. There are known cases where `ViewCreator` implementation holds reference to activity/fragment and that reference is leaked. Using hashcode of the object as key would work in the same way as using the object itself and we will get rid of potential memory leak issues.

Reviewed By: IanChilds

Differential Revision: D10510361

fbshipit-source-id: 351da8209afbb6ffd75124347f43c9f88874cfcd