Don't recursively calculate layout for released LayoutStateFutures
Summary:
The logic to handle released LayoutStateFutures can lead to infinite layout calculations.
If main thread is triggering a layout with layoutStateFuture1 and bg thread is triggering a new layout with layoutStateFuture2 before lsf1 finishes running, and lsf1 is not the same as lsf2, right now we'll invalidate lsf1 and start a new layout computation with lsf3. lsf3 is equal to lsf1 but not lsf2, and if lsf2 hasn't finished calculating, it will also be invalidated and trigger a new layout calculation. This will go on in a loop until the ongoing lsf can finish the calculation before the recursion is triggered.

This makes the following changes:
- keep a list of LayoutStateFutures that are currently in the process of calculating layout
- if a new layout is triggered and there's a LayoutStateFuture in the list already calculating that layout, increase a reference count on the lsf from the list
- when a lsf finishes calculating the layout and the threads waiting on the result are unblocked, decrease the reference count on the lsf
- when the reference count gets to 0 (all threads waiting on that result are unblocked), release the lsf and remove it from the list

@[1572450978:Andy] you mentioned there's already a ref count implemented for the layoutStateFuture but I couldn't find it, do you remember where it is?

Reviewed By: astreet

Differential Revision: D10509135

fbshipit-source-id: 17a2ded2300a25127d01a69b7cf1d575ab559ef3