Change ComponentContext to not extend Context
Summary: We don't want ComponentContext to extend Context - at one point it was just a thin wrapper but now it contains things such as the ComponentTree, so when people save off an object that takes the Context, we can leak the ComponentTree. Instead, you should just access the android Context using getAndroidContext from the ComponentContext.

Reviewed By: marco-cova

Differential Revision: D10868229

fbshipit-source-id: a09259e9d1db13892a2027c5c3ef176ab461be1a