Mark Drawable As Dirty On Parameters Changed
Summary:
When we set some parameters of CardClipDrawable, we do not mark the drawable as dirty. That is causing buildClippingCorners() not to run, and therefore incorrect corners are set on components when the drawable is getting re-mounted in a different position.

Now we mark the drawable as dirty in all set*() methods.

Reviewed By: IanChilds

Differential Revision: D12908718

fbshipit-source-id: 51130348a8d9a3f96439942a8c386402cba72617