Replace ArraySet with ArrayList in GraphBinding
Summary:
ArraySet has issues with its static cache (see summary in D8966285). Tldr; is that its static cache can be corruped in multi-threaded environment and next time any ArraySet instance is used it it could throw ClassCastException even if it is in single-threaded environment.

In this particular case it can be easily swapped with ArrayList as the contents are scanned linearly and no membership checks are performed.

Reviewed By: IanChilds

Differential Revision: D12942298

fbshipit-source-id: 4852b6bd392eee7b0f436e07e488d5bdf3a070a6