Deprecate com.facebook.litho.internal.ArraySet
Summary: ArraySet has issues with its static cache (see summary in D8966285). Tldr; is that its static cache can be corruped in multi-threaded environment and next time any ArraySet instance is used it it could throw ClassCastException even if it is in single-threaded environment. Deprecating it so that developers are aware of that issue.

Reviewed By: IanChilds

Differential Revision: D12943547

fbshipit-source-id: 4d1cb09f67e1a8cfd46cf69332887280567e705b