Replace ArraySet with ArrayList in TransitionManager
Summary:
ArraySet has issues with its static cache (see summary in D8966285). Tldr; is that its static cache can be corruped in multi-threaded environment and next time any ArraySet instance is used it it could throw ClassCastException even if it is in single-threaded environment.

In this particular case it can be swapped with ArrayList. Even though it has remove operation, the list will be very small (animations running on root component) so it's fine for linear scan as opposed to using HashSet which has logarithmic remove, but uses more memory.

Reviewed By: IanChilds

Differential Revision: D12942778

fbshipit-source-id: b310e56d0d96e9f656474b7af65ca58bcc8a2712