Replace ArraySet with HashSet in RenderState
Summary:
ArraySet has issues with its static cache (see summary in D8966285). Tldr; is that its static cache can be corruped in multi-threaded environment and next time any ArraySet instance is used it it could throw ClassCastException even if it is in single-threaded environment.

In this case ArraySet can be replaced by HashSet as it doesn't require ordering.

Reviewed By: IanChilds

Differential Revision: D12943830

fbshipit-source-id: b74cf1c40c6d476c23cdf39a0307ed235f7e27a3