Explicitly defining that providing LayoutHandlerFactory overrides ThreadPoolHandler in RecyclerBinder
Summary:
BG layout can be computed in one of three ways:
(1) default: we create `DefaultLayoutHandler` when no specific handler is passed
(2) `RecyclerBinderBuilder#layoutHandlerFactory` is provided: handler from this factory is used
(3) `RecyclerBinderBuilder#threadPoolConfig` is provided: ThreadPoolLayoutHandler is created from this config.

It wasn't clear what will happen when both (2) and (3) is provided. This diff makes sure that factory has precedence over the threadpoolconfig.

Reviewed By: marco-cova

Differential Revision: D12989265

fbshipit-source-id: 6f47d81332578c93b7f8da641f99e4847a53b08f