Change Litho's ComponentAccessibilityDelegate to work with newer versions of ExploreByTouchHelper
Summary:
This diff splits the ComponentAccessibilityDelegate into two different AccessibilityDelegates.  One (named VirtualViewsComponentAccessibilityDelegate) retains the full functionality of the original implementation, including supporting the creation of virtual child views, and the other is a much simpler version, only really passing along the standard accessibility methods to Litho's event dispatcher.

The reason for this change is that in v26 of the Android Support library the ExploreByTouchHelper class, which the original ComponentAccessibilityDelegate expanded, changed so that it forces any element that has it attached as a delegate to be focusable.  Since Litho was attaching this to every node that it deemed needed a View for accessibility, that meant that it forced every one of these Views to be focusable.  This caused a number of issues, such as breaking child description coalescing (if a focusable parent has unfocusable children, the parent should pull the text from them to use as its own description and not allow focus on the children), forcing focus on any element with an accessibility role set (such as all images), and more.

Our original solution was a quick and dirty fork of the ExploreByTouchHelper class which removed the lines which forced focus.  This fork however has made it difficult to keep our support library up to date.  More importantly, this reliance on a fork also made Litho's accessibility support in its open source incarnation not as functional as it was internally.

By making two delegates here, we can remove the fork in ExploreByTouchHelper since we will only be using a delegate that extends it when the view already would have needed to be focusable (when it has virtual children).

***Note:  I will remove the fork in the ExploreByTouchHelper class in a separate diff, as these changes work both with and without the fork (and therefore works with both older and newer versions of the Android support library).***

Reviewed By: IanChilds

Differential Revision: D12811196

fbshipit-source-id: ce6800f9a879e65014cdd3269a0b4d4446052ffc