Add ellipsize prop to TextInputSpec
Summary:
Declare a new Prop on TextInputSpec for ellipsis support. As with the previous EditTextSpec, the ellipsis behavior is a pass through to the underlying EditText native view, and relies on TextUtils.TruncateAt for behavior definition.

We do not perform additional processing on the ellipsis behavior, relying on mount / unmount to set the property on the native view.

Reviewed By: astreet

Differential Revision: D12875356

fbshipit-source-id: fc910abc0b65f1b095bfc73420558dcb1d8ecfd1