Implement should create layout for new size spec
Summary:
`onCreateLayoutWithSizeSpec` has the downside that if we need to remeasure a tree (either because of a yoga double-measure or because the developer manually needs to call measure multiple times in their product logic), we will fully recreate the tree (i.e. call `onCreateLayoutWithSizeSpec` again for the whole tree and throw away the old one).

Instead, we would like to be able to check with the layout spec whether the measure spec it was originally measured with generated a layout that is still compatible with the new spec. In that case we just need to remeasure instead of recreating the entire layout tree again.

Reviewed By: astreet

Differential Revision: D10163124

fbshipit-source-id: ff20bd5aa67680dd141510335ebda8025e16424d