Add text watchers to the TextInput.shouldUpdate() implementation
Summary:
After bringing TextWatcher from EditTextSpec, it became apparent we were missing the corresponding shouldUpdate check to avoid remounting the native TextInput and losing contextual data like cursor position.

Add the comparator for TextWatchers in should update. Because we can't compare TextWatcher implementations directly, we rely on the standard equality comparer on the list, as well as pointer comparisons for the values in it.

Reviewed By: pasqualeanatriello

Differential Revision: D13037757

fbshipit-source-id: 61e73ca9bdb06b5b1804e52285a955906d5ebd78