Make sure that we check that the component tree has been mounted before rendering the tooltip
Summary:
There are cases where we don't synchronously render a tooltip and instead render it after time delay.
In some cases we are attempting to render a tooltip view right after orientation change (T36694109).
At this time the component hasn't yet necissarily been mounted.

To make the showTooltip api more robust and continuing off the theme of this diff D13040701. We shouldn't make any assumptions about the validity of the component tree when users call showTooltip.
And so we should check if the component tree has even been mounted yet as well as if it has been released.

Reviewed By: denley

Differential Revision: D13091531

fbshipit-source-id: 37d52160f2f698a840af8d19141bb355c12088d4