introduce MountItem.LAYOUT_FLAG_PHANTOM flag
Summary: Introducing concept of "phantom" LayoutOutputs. A LayoutOutput, that not necessarily has to be mounted (this is actually up to `MountState`). The use case for now is: we are going to create phantom layout outputs, for things that normally don't need one, but may be animated. When they do get animated, only then the LayoutOutput would be mounted in form of `ComponentHost` view, that is going to be an animation handle (now we always create and mount a LayoutOutput in these cases)

Reviewed By: astreet

Differential Revision: D10508640

fbshipit-source-id: c6634561a2bb0bb7a6a657b3bd159f4febe53644