Treat tree props passed to setRoot* the same as the root component.
Summary:
TreeProps set in a SectionContext must be also set in the ComponentContext of all views under that section. This is done by storing those tree props in TreePropsWrappedRenderInfo and threading them through to `ComponentTree.setRoot(...)` so that we can use them in `ComponentTree.calculateLayoutState(...)`.

However, ComponentTree.calculateLayoutState is called in two different places:
  1) In ComponentTree.calculateLayout which is called by `Component.Tree.setRoot....`
  2) In ComponentTree.measure when android wants us to render something but we don't have a compatible layout already computed.

In #1, we pass the proper tree props to set, in #2, we have no access to these tree props so we don't set them.

This descrepancy means sometimes we might render a layout with treeprops from the section and sometimes we wouldn't.  If the layout of the hierarchy is complicated enough such that it's not complete by the time the Android framework asks us to measure, `ComponentTree.measure()` will calculate and use a layout without the section's treeprops set.

This approach takes Andy Street's suggestion to treat setting treeprops and setting root as the same atomic operation.  The new treeprops ivar will be passed to `ComponentTree.calculateLayout` in all places where a non-null `root` is passed.

Reviewed By: astreet

Differential Revision: D12948738

fbshipit-source-id: 930b733edfb7146b3773f786f0f00163d8ed9550