Change signature of UpdateState
Summary: Right now, we pass in the state container and the component, and then we use the component to access the state container. However, it's always the case that the state container is the component's so we don't need the component at all, just the state container that is passed into the method. This is good for future changes, where the state container won't actually be associated with the component.

Reviewed By: muraziz

Differential Revision: D13097086

fbshipit-source-id: 5f2b6279c3c301712d35a2bf9c3ab771f9b3c862