Keep last measured layout as ThreadLocal
Summary:
the reason we discarded lastMeasuredLayout when making a shallow copy is because if different components were using the same InternalNode on different threads, one might try to release it while the other one is still using it. (see D8397669)
Keeping as a ThreadLocal var should help with that. paveldudka would be great if you could help confirm this is fixing the problem you're seeing

Reviewed By: pasqualeanatriello

Differential Revision: D13102189

fbshipit-source-id: 012561d6f2468bf00f90b5f90af09a86945c153f