Only recalculate range on main thread
Summary: To reduce the likelihood of deadlocks, we really don't want to be acquiring the RecyclerBinder lock in the background if possible. It turns out, we don't need to to do this range calculation here anyway because we'll do it when we actually apply the batches on the main thread.

Reviewed By: mihaelao

Differential Revision: D13180053

fbshipit-source-id: 312500ae7d4a8e45d32d5937e50d7477222acdd1