Adds ComparableIntIdDrawable
Summary:
Adds ComparableIntIdDrawable which should be used for the most common cases on state dependent drawables.

A comparable implementation for drawables where an explicit int identifier is compared to test for equivalence in `isEquivalentTo(ComparableDrawable)`. Use this implementation for trivial cases where the number of drawables are discrete and finite. More formally create a function `f(x)` where `x` is the set of inputs used to determine the drawable and maps it to a unique int. An example:

```
onCreateLayout(ComponentContext c, State Day day) {

  Drawable drawable = getDrawableForDay(day);
  int id = day.ordinal(); // Day in an enum

  return Text.create(c)
      .text(day)
      .background(
         ComparableIntIdDrawable.create(drawable, id)
       )
      .build();
}
```

Reviewed By: astreet

Differential Revision: D13177364

fbshipit-source-id: f357ce0679f9cc820d55c287d587ac519c9b38b0