introduce TransitionIdMap
Summary: Introducing `TransitionIdMap` class, that maps `TransitionId`s to values. It keeps track of TransitionId types, thus allowing efficient look for keys of a particular type. This is important for `TransitionManager` that needs to map `TransitionId`s to `TransitionManager.AnimationState`s, and quickly find a value when creating transition for the given `com.facebook.litho.Transition.TransitionUnit`

Reviewed By: astreet

Differential Revision: D13275279

fbshipit-source-id: d1c9580754d87c0bf5d59807cb97d2eed487606b