Introduce delegateAllowRemeasure in Wrapper
Summary: Since there are existing cases where a component is measured and then passed to a Wrapper to change its common props, this is introducing a new Builder method on the Wrapper which allows an already measured component to be set as delegate. This makes a new copy of the delegate and  the internal node from the previous mesaurement will be thrown away to allow remeasurement. This also changes the soft error to an exception if a measured component is passed as delegate in debug mode

Reviewed By: astreet

Differential Revision: D13339768

fbshipit-source-id: e78ef934d5f71472a6b792fdb50168711d67f633