Add code generation for @CachedValue and @OnCalculateCachedValue
Summary: This allows CachedValues to be used for LayoutSpecs and MountSpecs. The code generation works as follows: it generates a static class that takes the CachedValue's dependencies and has `equals` and `hashCode` so that our HashMap of cached values can determine if we've already calculated the cached value for those inputs or not. Then it has a method `getX()` that will get the cached value for you, either from the cache or by generating a new one and adding it to the cache. The accessor method that we have is updated to use `getX()` for cached values. You can see how the generation works quite easily from the test files.

Reviewed By: astreet

Differential Revision: D13406657

fbshipit-source-id: 21a6d184fdb974697b197354550691aa21da3f69