Expose previous/next dataset from ChangeSet to Change
Summary:
The stack of diffs is to expose the data of each change from previous section to next section to OnDataRendered callback.

The stack of diffs is to expose a change list to OnDataRendered callback. The list contains the changes between previous section to next section, and each change has reference associated to the previous/next data.

There're few steps to achieve this.

1. Add new APIs in `Change.java`, that ChangeSet and pass data into it when creating changes.
2. Add new APIs in `ChangeSet.java`, that DiffSectionSpec can pass data into it when creating a ChangeSet.
3. Modify SingleComponentSectionSpec to provide the data in onCreateChangeSet() method.
4. Modify DataDiffSectionSpec to provide the data in onCreateChangeSet() method.
5. Provide firstVisibleIndex, lastVisibleIndex, and change list into OnDataRendered callback.

This diff adds two new arguments when create a Change object. `mPrevData` is the data before the change, and `mNextData` is after the change.

Reviewed By: mihaelao

Differential Revision: D13414836

fbshipit-source-id: c04cc4bf7b811e14bf9bb61accf2467f9c36d70f