Refactor DelegateMethodGenerator
Summary: In future diffs I want to have two ways to delegate to the spec - one directly and one via reflection to enable hotswap to occur. Therefore, I'm refactoring the code a little here to make that easier to do in future diffs - we collect up all the types and param names and then add them at the end, rather than doing it as we go. This does lead to some extra (unnecessary) casting, but I don't see that as a problem.

Reviewed By: passy

Differential Revision: D13521780

fbshipit-source-id: f57d84062d5eb2d8ceb193f3b445aefaf43c50e1